requirements.txt
langchain-google-vertexai
langchain-huggingface
langchain-community
langchain-core
langchain-text-splitters
chromadb
elasticsearch
sentence-transformers
torch
transformers
tavily-python
pydantic
numpy
pymupdf
json5
requests
beautifulsoup4
datasets
